+++ modules/tinymce/src/core/main/ts/content/ContentTypes.ts
@@ -5,13 +5,16 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
+import Editor from '../api/Editor';
 import AstNode from '../api/html/Node';
 
 export type Content = string | AstNode;
-export type ContentFormat = 'raw' | 'text' | 'html' | 'tree';
+
+export type GetContentFormatter = (editor: Editor, args: GetContentArgs) => Content;
+export type SetContentFormatter = (editor: Editor, content: Content, args: SetContentArgs) => Content;
 
 export interface GetContentArgs {
-  format?: ContentFormat;
+  format?: string;
   get?: boolean;
   content?: string;
   getInner?: boolean;