+++ modules/tinymce/src/core/main/ts/selection/GetSelectionContent.ts
@@ -6,12 +6,12 @@
  */
 
 import Editor from '../api/Editor';
-import { Content, ContentFormat } from '../content/ContentTypes';
+import { Content } from '../content/ContentTypes';
 import * as Rtc from '../Rtc';
 import { GetSelectionContentArgs } from './GetSelectionContentImpl';
 
 const getContent = (editor: Editor, args: GetSelectionContentArgs = {}): Content => {
-  const format: ContentFormat = args.format ? args.format : 'html';
+  const format: string = args.format ? args.format : 'html';
 
   return Rtc.getSelectedContent(editor, format, args);
 };